Frequencia em Ranking

+++ src/SME.SGP.Dominio/Entidades/PainelEducacionalRegistroFrequenciaAgrupamentoMensal.cs
@@ -3,7 +3,7 @@
     public class PainelEducacionalRegistroFrequenciaAgrupamentoMensal : EntidadeBase
     {
         public int Modalidade { get; set; }
-        public long AnoLetivo { get; set; }
+        public int AnoLetivo { get; set; }
         public int Mes { get; set; }
         public int TotalAulas { get; set; }
         public int TotalFaltas { get; set; }
